Got rid of warnings in CSM - all of om build for me without warnings now

+++ trunk/openmodeller/src/algorithms/csm.cpp
@@ -375,7 +375,7 @@ Scalar Csm::getValue( const Sample& x ) const
   if (verboseDebuggingBool)
   {
     printf("\n-------------------------------\n");
-    printf("Component count : %u\n",z->size2);
+    printf("Component count : %u\n",static_cast<unsigned int>(z->size2));
     printf("Component count / 2: %f\n",myHalfComponentCountDouble);
     printf("Sum of squares : %f\n",mySumOfSquares);
     printf("Sum of squares / 2: %f\n",myHalfSumOfSquaresDouble);
@@ -416,11 +416,11 @@ void Csm::displayVector(const gsl_vector * v, const char * name, const bool roun
     if (roundFlag)
     {
 
-      fprintf( stderr, "\nDisplaying Vector rounded to 4 decimal places '%s' (%u): \n----------------------------------------------\n[  ", name, v->size );
+      fprintf( stderr, "\nDisplaying Vector rounded to 4 decimal places '%s' (%u): \n----------------------------------------------\n[  ", name, static_cast<unsigned int>(v->size) );
     }
     else
     {
-      fprintf( stderr, "\nDisplaying Vector '%s' (%u): \n----------------------------------------------\n[  ", name, v->size );
+      fprintf( stderr, "\nDisplaying Vector '%s' (%u): \n----------------------------------------------\n[  ", name, static_cast<unsigned int>(v->size) );
     }
 
     char sep1[] = ", ";
@@ -453,11 +453,11 @@ void Csm::displayMatrix(const gsl_matrix * m, const char * name, const bool roun
   {
     if (!roundFlag)
     {
-      fprintf( stderr, "\nDisplaying Matrix '%s' (%u / %u): \n----------------------------------------------\n[\n", name, m->size1, m->size2 );
+      fprintf( stderr, "\nDisplaying Matrix '%s' (%u / %u): \n----------------------------------------------\n[\n", name, static_cast<unsigned int>(m->size1), static_cast<unsigned int>(m->size2) );
     }
     else
     {
-      fprintf( stderr, "\nDisplaying Matrix rounded to 4 decimal places '%s' (%u / %u): \n----------------------------------------------\n[\n", name, m->size1, m->size2 );
+      fprintf( stderr, "\nDisplaying Matrix rounded to 4 decimal places '%s' (%u / %u): \n----------------------------------------------\n[\n", name, static_cast<unsigned int>(m->size1), static_cast<unsigned int>(m->size2) );
     }
     for (unsigned int i=0;i<m->size1;++i)
     {
